Refactor import path for get_trial_filtering_prompt and add new trial filter prompt template
